+++ src/test/java/io/cryostat/discovery/DiscoveryTest.java
@@ -17,7 +17,6 @@ package io.cryostat.discovery;
 
 import static io.restassured.RestAssured.given;
 
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
@@ -28,14 +27,7 @@ import io.quarkus.test.junit.QuarkusTest;
 import io.restassured.http.ContentType;
 import org.hamcrest.MatcherAssert;
 import org.hamcrest.Matchers;
-import org.junit.jupiter.api.MethodOrderer.OrderAnnotation;
-import org.junit.jupiter.api.Nested;
-import org.junit.jupiter.api.Order;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.TestMethodOrder;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.NullAndEmptySource;
-import org.junit.jupiter.params.provider.ValueSource;
 
 @QuarkusTest
 @TestHTTPEndpoint(Discovery.class)
@@ -93,66 +85,6 @@ public class DiscoveryTest extends AbstractTransactionalTestBase {
                     plugin,
                     Matchers.hasEntry(
                             Matchers.equalTo("id"), Matchers.not(Matchers.blankOrNullString())));
-        }
-    }
-
-    @Nested
-    class PluginValidations {
-        @ParameterizedTest
-        @NullAndEmptySource
-        @ValueSource(strings = {"invalid uri", "no.protocol.example.com"})
-        void rejectsInvalidCallback(String callback) {
-            Map<String, String> payload = new HashMap<>();
-            payload.put("callback", callback);
-            payload.put("realm", "test");
-            given().log()
-                    .all()
-                    .when()
-                    .body(payload)
-                    .contentType(ContentType.JSON)
-                    .post("/api/v4/discovery")
-                    .then()
-                    .log()
-                    .all()
-                    .and()
-                    .assertThat()
-                    .statusCode(400);
-        }
-
-        @ParameterizedTest
-        @NullAndEmptySource
-        void rejectsInvalidRealmName(String realm) {
-            Map<String, String> payload = new HashMap<>();
-            payload.put("callback", "http://example.com");
-            payload.put("realm", realm);
-            given().log()
-                    .all()
-                    .when()
-                    .body(payload)
-                    .contentType(ContentType.JSON)
-                    .post("/api/v4/discovery")
-                    .then()
-                    .log()
-                    .all()
-                    .and()
-                    .assertThat()
-                    .statusCode(400);
-        }
-
-        @Test
-        void rejectsPublishForUnregisteredPlugin() {
-            given().log()
-                    .all()
-                    .when()
-                    .body(List.of())
-                    .contentType(ContentType.JSON)
-                    .post("/api/v4/discovery/abcd1234")
-                    .then()
-                    .log()
-                    .all()
-                    .and()
-                    .assertThat()
-                    .statusCode(404);
         }
     }
 }
